requirements.txt
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
flask==2.3.3
greenlet==3.0.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.0
outcome==1.2.0
pandas==2.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.21
trio==0.22.2
trio-websocket==0.11.1
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
werkzeug==2.3.7
wsproto==1.2.0
zipp==3.17.0
